Postgres search refactor + fixes, part 1 (#875)
Handles most types now, only slices of enums are left.

+++ central/processbaselineresults/datastore/internal/store/postgres/store.go
@@ -91,7 +91,7 @@ func createTableProcesswhitelistresultsBaselineStatuses(ctx context.Context, db 
 	table := `
 create table if not exists processwhitelistresults_BaselineStatuses (
     processwhitelistresults_DeploymentId varchar,
-    idx numeric,
+    idx integer,
     ContainerName varchar,
     BaselineStatus integer,
     AnomalousProcessesExecuted bool,
